Initial Landing Documentation with navigations to different trainings. Added Navigation links at the end of each of the k8s Training documentation. Fixed the GKE Cluster setup.sh script for initialization issue by externalization of GKE cluster version as input. Added validation logic to handle the deletion without error in the teardown.sh script. Cleaned up unwanted documentation related to setup.

+++ k8s_fundamentals/99_teardown/teardown_cluster.sh
@@ -22,9 +22,62 @@ gcloud config set compute/region $REGION
 gcloud config set compute/zone $ZONE
 
 # delete resources
-gcloud beta container clusters delete $CLUSTER_NAME --quiet
-gcloud compute firewall-rules delete $FIREWALL_NAME-ingress-gateway --quiet
-gcloud compute firewall-rules delete $FIREWALL_NAME-ssh --quiet
-gcloud compute firewall-rules delete $FIREWALL_NAME-nodeport --quiet
-gcloud compute networks subnets delete $NETWORK_NAME-subnet --quiet
-gcloud compute networks delete $NETWORK_NAME --quiet
+#Check whether Cluster exists
+response_cluster=`gcloud beta container clusters list --filter="name=$CLUSTER_NAME" --format="value(name)" --project=$PROJECT_NAME`
+
+if [ ! -z "$response_cluster" ]
+then
+  gcloud beta container clusters delete $CLUSTER_NAME --quiet
+else
+  echo "Cluster $response_cluster doesn't exists, skipped deletion"
+fi
+
+# Check whether the Firewall Rules for Ingress Gateway exists
+response_firewall_ingress=`gcloud compute firewall-rules list --filter="$FIREWALL_NAME-ingress-gateway" --format="value(name)" --project=$PROJECT_NAME`
+
+if [ ! -z "$response_firewall_ingress" ]
+then
+  gcloud compute firewall-rules delete $FIREWALL_NAME-ingress-gateway --quiet
+else
+  echo "Firewall Rules for Ingress Gateway $response_firewall_ingress doesn't exists, skipped deletion"
+fi
+
+# Check whether the Firewall Rules for SSH Port exists
+response_firewall_ssh=`gcloud compute firewall-rules list --filter="$FIREWALL_NAME-ssh" --format="value(name)" --project=$PROJECT_NAME`
+
+if [ ! -z "$response_firewall_ssh" ]
+then
+  gcloud compute firewall-rules delete $FIREWALL_NAME-ssh --quiet
+else
+  echo "Firewall Rules for SSH $response_firewall_ssh doesn't exists, skipped deletion"
+fi
+
+# Check whether the Firewall Rules for NodePort access exists
+response_firewall_nodeport=`gcloud compute firewall-rules list --filter="$FIREWALL_NAME-nodeport" --format="value(name)" --project=$PROJECT_NAME`
+
+if [ ! -z "$response_firewall_nodeport" ]
+then
+  gcloud compute firewall-rules delete $FIREWALL_NAME-nodeport --quiet
+else
+  echo "Firewall Rules for NodePort $response_firewall_nodeport doesn't exists, skipped deletion"
+fi
+
+#Check whether Cluster Subnet exists
+response_subnet=`gcloud compute networks subnets list --filter="name=$NETWORK_NAME-subnet" --format="value(name)" --project=$PROJECT_NAME`
+
+if [ ! -z "$response_subnet" ]
+then
+  gcloud compute networks subnets delete $NETWORK_NAME-subnet --quiet
+else
+  echo "Subnet $response_subnet doesn't exists, skipped deletion"
+fi
+
+#Check whether Cluster Network exists
+response_network=`gcloud compute networks list --filter="name=$NETWORK_NAME" --format="value(name)" --project=$PROJECT_NAME`
+
+if [ ! -z "$response_network" ]
+then
+  gcloud compute networks delete $NETWORK_NAME --quiet
+else
+  echo "Network $response_network doesn't exists, skipped deletion"
+fi
